+++ src/main/java/insane96mcp/iguanatweaksexpanded/data/generator/ITEDamageTypeTagsProvider.java
@@ -1,8 +1,8 @@
 package insane96mcp.iguanatweaksexpanded.data.generator;
 
 import insane96mcp.iguanatweaksexpanded.IguanaTweaksExpanded;
-import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment.BloodPact;
 import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment.MagicProtection;
+import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment.curse.CurseOfBloodPact;
 import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment.curse.CurseOfExperience;
 import insane96mcp.iguanatweaksexpanded.module.items.copper.CopperExpansion;
 import insane96mcp.iguanatweaksreborn.module.combat.PiercingDamage;
@@ -32,8 +32,8 @@ public class ITEDamageTypeTagsProvider extends DamageTypeTagsProvider {
 
         tag(CopperExpansion.DOESNT_TRIGGER_ELECTROCUTION).addOptionalTag(PiercingDamage.PIERCING_DAMAGE_TYPE.location()).add(CopperExpansion.ELECTROCUTION_ATTACK).add(CurseOfExperience.DAMAGE_TYPE);
 
-        tag(DamageTypeTags.BYPASSES_ARMOR).add(CurseOfExperience.DAMAGE_TYPE, BloodPact.DAMAGE_TYPE);
-        tag(DamageTypeTags.BYPASSES_COOLDOWN).add(BloodPact.DAMAGE_TYPE);
+        tag(DamageTypeTags.BYPASSES_ARMOR).add(CurseOfExperience.DAMAGE_TYPE, CurseOfBloodPact.DAMAGE_TYPE);
+        tag(DamageTypeTags.BYPASSES_COOLDOWN).add(CurseOfBloodPact.DAMAGE_TYPE);
 
         tag(MagicProtection.SOURCES_REDUCED).add(DamageTypes.MAGIC, DamageTypes.INDIRECT_MAGIC);
     }