+++ changelog.md
@@ -6,6 +6,8 @@
   * Reduced Perfect Parry cost
   * Solarium shields now give +3 blocked damage in the sun (was 2)
   * Fixed copper shields not having bonus unbreaking when used underground
+* Blood Pact is now Curse of Blood Pact
+  * +6 max enchanting cost the user has 1/10 chance to take 1 true damage whenever the tool takes damage
 * Soul Blast Furnaces current fuel no longer consumes if not smelting anything
 * Reduced Protection enchantment cost (5 -> 4)
 * Spiders now attack much slower (like enderman)